Add files via upload
umarım gene doğru dosyayı yüklemişimdir 
variable jump yapmaya çalıştım
kamera geçişlerini yapmaya çalıştım odadan odaya olan (biraz buglı ama düzeltemedim belki karakter spawnı yüzünden olabilir)
düşmanalar bizi görünce takip etsin saldırsın boş boş dolanmasınlar diye onu ekledim
ek olarak bize saldırmaları için ana karakterin vuruş efektine benzer ama daha gecikmeli ekledim(buda bug lı ilk vuruşu çok ani atıyorlar ek olarak aynı anda falan vurunca tek vuruşa ölüyorlar sanki parry gibi oldu )

+++ KeserKnight/KeserKnight/Core/PhysicsEngine.cs
@@ -31,11 +31,11 @@ namespace KeserKnight.Core
                 else nextX += player.Speed;
             }
 
-            int nextY = player.Y + player.VerticalVelocity;
+            float nextY = player.Y + player.VerticalVelocity;
 
             // 3. Karakteri doğrudan hedef koordinatına taşı usta (Çift adımdan kurtulduk)
             player.X = nextX;
-            player.Y = nextY;
+            player.Y = (int)nextY;
 
             // Havada kalma durumunu başta aktif et, eğer zemine basarsa aşağıda false yapacağız
             player.IsJumping = true;
